refactor convert to int helper
