+++ protocol/reseed/reseed7.js
@@ -26,15 +26,3 @@ async function reseed7(account, L2Beanstalk) {
   console.log("-----------------------------------");
 }
 exports.reseed7 = reseed7;
-
-// Helper function to recursively convert string numbers to integers
-function convertToInt(value) {
-  // Check if the value is a valid address format
-  const isAddress = /^0x[a-fA-F0-9]/.test(value);
-  if (Array.isArray(value)) {
-    return value.map(convertToInt);
-  } else if (typeof value === 'string' && !isAddress && !isNaN(value)) {
-    return parseInt(value, 10);
-  }
-  return value;
-}
