+++ protocol/utils/read.js
@@ -7,4 +7,17 @@ async function readPrune() {
     return num
 }
 
+// Helper function to recursively convert string numbers to integers
+function convertToInt(value) {
+    // Check if the value is a valid address format
+    const isAddress = /^0x[a-fA-F0-9]/.test(value);
+    if (Array.isArray(value)) {
+      return value.map(convertToInt);
+    } else if (typeof value === 'string' && !isAddress && !isNaN(value)) {
+      return parseInt(value, 10);
+    }
+    return value;
+  }
+
 exports.readPrune = readPrune
+exports.convertToInt = convertToInt
